Add files via upload
When you attach a file it is copied to the "Uploads Folder". Now need a way to offer user option to download. (I was thinking of a hyperlink)
When Attaching an image only images will pop (gif, jpg, png and tiff).
Number of online users works but but doesnt refresh, on subsequent open windows the accurate number is display